// src/interfaces/IConfiguration.ts
export interface IConfiguration {
  MS_GRAPH_APP_ID: string;
  MS_GRAPH_DIRECTORY_ID: string;
  AZURE_DEVOPS_ORGANISATION: string;
  AZURE_DEVOPS_API_TIMEOUT: number;
  AZURE_DEVOPS_API_MAX_RETRIES: number;
  AZURE_DEVOPS_API_ALLOWRETRIES: boolean;
  MS_GRAPH_TOKEN_ENDPOINT: string;
  MS_GRAPH_TOKEN_POSTFIX: string;
  MS_GRAPH_SCOPE: string;
  AZURE_DEVOPS_ORGANISATION_URL: string;
  AZURE_DEVOPS_ORGANISATION_VSSPS_URL: string;
  AZURE_DEVOPS_ORGANISATION_VSAEX_URL: string;
  TESTING: boolean;
  TEAM_PROJECT_PROCESS_TEMPLATE_ID: string;
  MAX_LOOP_FOR_PROJECT_QUEING: number;
  AAD_TEAM_GROUP_PREFIX: string;
  AAD_PROJECT_READER_GROUP_PREFIX: string;
  AAD_PRODUCT_GROUP_PREFIX: string;
  AZURE_DEVOPS_INNERSOURCE_TEAM_GROUP_PREFIX: string;
  AZURE_DEVOPS_INNERSOURCE_PRODUCT_GROUP_PREFIX: string;
  AZURE_DEVOPS_SECURITY_GROUP_PREFIX: string;
  AZURE_DEVOPS_TEAM_GROUP_PREFIX: string;
  AZURE_DEVOPS_PRODUCT_GROUP_PREFIX: string;
  AZURE_DEVOPS_VALUESTREAM_TEAM_PREFIX: string;
  AZURE_DEVOPS_API_MAX_ITEMS_TO_RETREIVE: string;
  CONFIG_GITPOLICYFILE: string;
  CONFIG_PROJECTPERMISSIONFILE: string;
  CONFIG_SIMPLEOBJECTPERMISSIONFILE: string;
  CONFIG_BUILDFOLERPERMISSIONFILE: string;
  CONFIG_RELEASEFOLERPERMISSIONFILE: string;
  CONFIG_PRODUCTVARIABLEGROUPFILE: string;
  CONFIG_PROJECTVARIABLEGROUPFILE: string;
  CONFIG_GITPERMISSIONFILE: string;
  DEFAULTAGENTQUEUES: string;
  AZUREDEVOPS_PAT?: string;
  AZUREDEVOPS_PAT_OWNER?: string;
  MS_GRAPH_APP_SECRET?: string;
}